Changed around how the startup of prison works. This helps to eliminate a couple of error messages when enabling the compatibility classes. Moved the dispaly of the system stats to the Prison class so it can be now included with the prison splash screen.  This is important to get displayed here so if there are other failures, at least we have that covered to help with troubleshooting.
